remove processMerge
